Linker is missing on ubuntu, fixing it

+++ builder/__main__.py
@@ -215,7 +215,8 @@ if __name__ == '__main__':
     validate_build(env.build_spec)
 
     # Once initialized, switch to the source dir before running actions
-    env.shell.rm(env.build_dir)
+    if os.path.exists(env.build_dir):
+        env.shell.rm(env.build_dir)
     env.shell.mkdir(env.build_dir)
     env.shell.cd(env.project.path)
